tools/requirements-autotest.txt
pytest>=8.0.0
requests>=2.31.0
python-dotenv>=1.0.0

